backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.8.0
openai>=1.40.0
datasets>=2.20.0
huggingface_hub>=0.24.0
sdv>=1.17.0
python-dotenv>=1.0.1
